requirements.txt
streamlit>=1.30.0
anthropic>=0.18.0
openai>=1.0.0
google-generativeai>=0.8.0
fpdf2>=2.8.0
